pyproject.toml
[project]
name = "ubongo-iq"
version = "0.1.0"
description = ""
authors = [
    {name = "Mwapsam",email = "mwapsam@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "wagtail (>=7.0.1,<8.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "django-storages (>=1.14.6,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "mailtrap (>=2.1.0,<3.0.0)",
    "boto3 (>=1.39.4,<2.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-compressor (>=4.5.1,<5.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pillow-avif-plugin (>=1.5.2,<2.0.0)",
    "readtime (>=3.0.0,<4.0.0)",
    "sentry-sdk[django] (>=2.34.1,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "selenium (>=4.35.0,<5.0.0)"
]

[tool.poetry]
packages = [{include = "ubongo_iq", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.11.1"
webdriver-manager = "^4.0.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
